+++ py2app/recipes/__init__.py
@@ -19,6 +19,7 @@ from . import qt5  # noqa: F401
 from . import re  # noqa: F401
 from . import scipy  # noqa: F401
 from . import sip  # noqa: F401
+from . import six  # noqa: F401
 from . import subprocess  # noqa: F401
 from . import sysconfig_module  # noqa: F401
 from . import tkinter  # noqa: F401